Katello distributions -> Foreman installation media integration
Create installation media for every http published repo with distribution
in Katello. Also create architecture and operating system with sane defaults if
missing in Foreman.

When a distribution is removed from the repository, or the repository is demoted,
the installation media is removed.

Both default environments and content views are supported.